@@ -15,17 +15,15 @@ limitations under the License.
 */
 
 // TODO: Ensure we have proper error handling/logging for API calls to Bluecat. getBluecatGatewayToken has a good example of this
+// TODO: Remove studdering
+// TODO: Make API calls more consistent (eg error handling on HTTP response codes)
+// TODO: zone-id-filter does not seem to work with our provider
 
 package bluecat
 
 import (
-	"bytes"
 	"context"
-	"crypto/tls"
 	"encoding/json"
-	"io"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"regexp"
 	"strconv"
@@ -37,17 +35,8 @@ import (
 	"sigs.k8s.io/external-dns/endpoint"
 	"sigs.k8s.io/external-dns/plan"
 	"sigs.k8s.io/external-dns/provider"
+	api "sigs.k8s.io/external-dns/provider/bluecat/gateway"
 )
-
-type bluecatConfig struct {
-	GatewayHost      string `json:"gatewayHost"`
-	GatewayUsername  string `json:"gatewayUsername,omitempty"`
-	GatewayPassword  string `json:"gatewayPassword,omitempty"`
-	DNSConfiguration string `json:"dnsConfiguration"`
-	View             string `json:"dnsView"`
-	RootZone         string `json:"rootZone"`
-	SkipTLSVerify    bool   `json:"skipTLSVerify"`
-}
 
 // BluecatProvider implements the DNS provider for Bluecat DNS
 type BluecatProvider struct {
@@ -57,67 +46,12 @@ type BluecatProvider struct {
 	dryRun           bool
 	RootZone         string
 	DNSConfiguration string
+	DNSServerName    string
+	DNSDeployType    string
 	View             string
-	gatewayClient    GatewayClient
-}
-
-type GatewayClient interface {
-	getBluecatZones(zoneName string) ([]BluecatZone, error)
-	getHostRecords(zone string, records *[]BluecatHostRecord) error
-	getCNAMERecords(zone string, records *[]BluecatCNAMERecord) error
-	getHostRecord(name string, record *BluecatHostRecord) error
-	getCNAMERecord(name string, record *BluecatCNAMERecord) error
-	createHostRecord(zone string, req *bluecatCreateHostRecordRequest) (res interface{}, err error)
-	createCNAMERecord(zone string, req *bluecatCreateCNAMERecordRequest) (res interface{}, err error)
-	deleteHostRecord(name string, zone string) (err error)
-	deleteCNAMERecord(name string, zone string) (err error)
-	buildHTTPRequest(method, url string, body io.Reader) (*http.Request, error)
-	getTXTRecords(zone string, records *[]BluecatTXTRecord) error
-	getTXTRecord(name string, record *BluecatTXTRecord) error
-	createTXTRecord(zone string, req *bluecatCreateTXTRecordRequest) (res interface{}, err error)
-	deleteTXTRecord(name string, zone string) error
-}
-
-// GatewayClientConfig defines new client on bluecat gateway
-type GatewayClientConfig struct {
-	Cookie           http.Cookie
-	Token            string
-	Host             string
-	DNSConfiguration string
-	View             string
-	RootZone         string
-	SkipTLSVerify    bool
-}
-
-// BluecatZone defines a zone to hold records
-type BluecatZone struct {
-	ID         int    `json:"id"`
-	Name       string `json:"name"`
-	Properties string `json:"properties"`
-	Type       string `json:"type"`
-}
-
-// BluecatHostRecord defines dns Host record
-type BluecatHostRecord struct {
-	ID         int    `json:"id"`
-	Name       string `json:"name"`
-	Properties string `json:"properties"`
-	Type       string `json:"type"`
-}
-
-// BluecatCNAMERecord defines dns CNAME record
-type BluecatCNAMERecord struct {
-	ID         int    `json:"id"`
-	Name       string `json:"name"`
-	Properties string `json:"properties"`
-	Type       string `json:"type"`
-}
-
-// BluecatTXTRecord defines dns TXT record
-type BluecatTXTRecord struct {
-	ID         int    `json:"id"`
-	Name       string `json:"name"`
-	Properties string `json:"properties"`
+	gatewayClient    api.GatewayClient
+	TxtPrefix        string
+	TxtSuffix        string
 }
 
 type bluecatRecordSet struct {
@@ -125,45 +59,44 @@ type bluecatRecordSet struct {
 	res interface{}
 }
 
-type bluecatCreateHostRecordRequest struct {
-	AbsoluteName string `json:"absolute_name"`
-	IP4Address   string `json:"ip4_address"`
-	TTL          int    `json:"ttl"`
-	Properties   string `json:"properties"`
-}
-
-type bluecatCreateCNAMERecordRequest struct {
-	AbsoluteName string `json:"absolute_name"`
-	LinkedRecord string `json:"linked_record"`
-	TTL          int    `json:"ttl"`
-	Properties   string `json:"properties"`
-}
-
-type bluecatCreateTXTRecordRequest struct {
-	AbsoluteName string `json:"absolute_name"`
-	Text         string `json:"txt"`
-}
-
 // NewBluecatProvider creates a new Bluecat provider.
 //
 // Returns a pointer to the provider or an error if a provider could not be created.
-func NewBluecatProvider(configFile string, domainFilter endpoint.DomainFilter, zoneIDFilter provider.ZoneIDFilter, dryRun bool) (*BluecatProvider, error) {
-	contents, err := ioutil.ReadFile(configFile)
+func NewBluecatProvider(configFile, dnsConfiguration, dnsServerName, dnsDeployType, dnsView, gatewayHost, rootZone, txtPrefix, txtSuffix string, domainFilter endpoint.DomainFilter, zoneIDFilter provider.ZoneIDFilter, dryRun, skipTLSVerify bool) (*BluecatProvider, error) {
+	cfg := api.BluecatConfig{}
+	contents, err := os.ReadFile(configFile)
 	if err != nil {
-		return nil, errors.Wrapf(err, "failed to read Bluecat config file %v", configFile)
+		if errors.Is(err, os.ErrNotExist) {
+			cfg = api.BluecatConfig{
+				GatewayHost:      gatewayHost,
+				DNSConfiguration: dnsConfiguration,
+				DNSServerName:    dnsServerName,
+				DNSDeployType:    dnsDeployType,
+				View:             dnsView,
+				RootZone:         rootZone,
+				SkipTLSVerify:    skipTLSVerify,
+				GatewayUsername:  "",
+				GatewayPassword:  "",
+			}
+		} else {
+			return nil, errors.Wrapf(err, "failed to read Bluecat config file %v", configFile)
+		}
+	} else {
+		err = json.Unmarshal(contents, &cfg)
+		if err != nil {
+			return nil, errors.Wrapf(err, "failed to parse Bluecat JSON config file %v", configFile)
+		}
 	}
 
-	cfg := bluecatConfig{}
-	err = json.Unmarshal(contents, &cfg)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to read Bluecat config file %v", configFile)
+	if !api.IsValidDNSDeployType(cfg.DNSDeployType) {
+		return nil, errors.Errorf("%v is not a valid deployment type", cfg.DNSDeployType)
 	}
 
-	token, cookie, err := getBluecatGatewayToken(cfg)
+	token, cookie, err := api.GetBluecatGatewayToken(cfg)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to get API token from Bluecat Gateway")
 	}
-	gatewayClient := NewGatewayClient(cookie, token, cfg.GatewayHost, cfg.DNSConfiguration, cfg.View, cfg.RootZone, cfg.SkipTLSVerify)
+	gatewayClient := api.NewGatewayClientConfig(cookie, token, cfg.GatewayHost, cfg.DNSConfiguration, cfg.View, cfg.RootZone, cfg.DNSServerName, cfg.SkipTLSVerify)
 
 	provider := &BluecatProvider{
 		domainFilter:     domainFilter,
@@ -171,28 +104,14 @@ func NewBluecatProvider(configFile string, domainFilter endpoint.DomainFilter, z
 		dryRun:           dryRun,
 		gatewayClient:    gatewayClient,
 		DNSConfiguration: cfg.DNSConfiguration,
+		DNSServerName:    cfg.DNSServerName,
+		DNSDeployType:    cfg.DNSDeployType,
 		View:             cfg.View,
 		RootZone:         cfg.RootZone,
+		TxtPrefix:        txtPrefix,
+		TxtSuffix:        txtSuffix,
 	}
 	return provider, nil
-}
-
-// NewGatewayClient creates and returns a new Bluecat gateway client
-func NewGatewayClient(cookie http.Cookie, token, gatewayHost, dnsConfiguration, view, rootZone string, skipTLSVerify bool) GatewayClientConfig {
-	// Right now the Bluecat gateway doesn't seem to have a way to get the root zone from the API. If the user
-	// doesn't provide one via the config file we'll assume it's 'com'
-	if rootZone == "" {
-		rootZone = "com"
-	}
-	return GatewayClientConfig{
-		Cookie:           cookie,
-		Token:            token,
-		Host:             gatewayHost,
-		DNSConfiguration: dnsConfiguration,
-		View:             view,
-		RootZone:         rootZone,
-		SkipTLSVerify:    skipTLSVerify,
-	}
 }
 
 // Records fetches Host, CNAME, and TXT records from bluecat gateway
@@ -206,8 +125,8 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 	for _, zone := range zones {
 		log.Debugf("fetching records from zone '%s'", zone)
 
-		var resT []BluecatTXTRecord
-		err = p.gatewayClient.getTXTRecords(zone, &resT)
+		var resT []api.BluecatTXTRecord
+		err = p.gatewayClient.GetTXTRecords(zone, &resT)
 		if err != nil {
 			return nil, errors.Wrapf(err, "could not fetch TXT records for zone: %v", zone)
 		}
@@ -220,14 +139,14 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 			endpoints = append(endpoints, tempEndpoint)
 		}
 
-		var resH []BluecatHostRecord
-		err = p.gatewayClient.getHostRecords(zone, &resH)
+		var resH []api.BluecatHostRecord
+		err = p.gatewayClient.GetHostRecords(zone, &resH)
 		if err != nil {
 			return nil, errors.Wrapf(err, "could not fetch host records for zone: %v", zone)
 		}
 		var ep *endpoint.Endpoint
 		for _, rec := range resH {
-			propMap := splitProperties(rec.Properties)
+			propMap := api.SplitProperties(rec.Properties)
 			ips := strings.Split(propMap["addresses"], ",")
 			for _, ip := range ips {
 				if _, ok := propMap["ttl"]; ok {
@@ -240,7 +159,7 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 					ep = endpoint.NewEndpoint(propMap["absoluteName"], endpoint.RecordTypeA, ip)
 				}
 				for _, txtRec := range resT {
-					if strings.Compare(rec.Name, txtRec.Name) == 0 {
+					if strings.Compare(p.TxtPrefix+rec.Name+p.TxtSuffix, txtRec.Name) == 0 {
 						ep.Labels[endpoint.OwnerLabelKey], err = extractOwnerfromTXTRecord(txtRec.Properties)
 						if err != nil {
 							log.Debugf("External DNS Owner %s", err)
@@ -251,14 +170,14 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 			}
 		}
 
-		var resC []BluecatCNAMERecord
-		err = p.gatewayClient.getCNAMERecords(zone, &resC)
+		var resC []api.BluecatCNAMERecord
+		err = p.gatewayClient.GetCNAMERecords(zone, &resC)
 		if err != nil {
 			return nil, errors.Wrapf(err, "could not fetch CNAME records for zone: %v", zone)
 		}
 
 		for _, rec := range resC {
-			propMap := splitProperties(rec.Properties)
+			propMap := api.SplitProperties(rec.Properties)
 			if _, ok := propMap["ttl"]; ok {
 				ttl, err := strconv.Atoi(propMap["ttl"])
 				if err != nil {
@@ -269,7 +188,7 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 				ep = endpoint.NewEndpoint(propMap["absoluteName"], endpoint.RecordTypeCNAME, propMap["linkedRecordName"])
 			}
 			for _, txtRec := range resT {
-				if strings.Compare(rec.Name, txtRec.Name) == 0 {
+				if strings.Compare(p.TxtPrefix+rec.Name+p.TxtSuffix, txtRec.Name) == 0 {
 					ep.Labels[endpoint.OwnerLabelKey], err = extractOwnerfromTXTRecord(txtRec.Properties)
 					if err != nil {
 						log.Debugf("External DNS Owner %s", err)
@@ -278,8 +197,6 @@ func (p *BluecatProvider) Records(ctx context.Context) (endpoints []*endpoint.En
 			}
 			endpoints = append(endpoints, ep)
 		}
-
-		// TODO: add bluecat deploy API call here
 	}
 
 	log.Debugf("fetched %d records from Bluecat", len(endpoints))
@@ -301,6 +218,24 @@ func (p *BluecatProvider) ApplyChanges(ctx context.Context, changes *plan.Change
 	log.Infof("deleted: %+v\n", deleted)
 	p.deleteRecords(deleted)
 	p.createRecords(created)
+
+	if p.DNSServerName != "" {
+		if p.dryRun {
+			log.Debug("Not executing deploy because this is running in dry-run mode")
+		} else {
+			switch p.DNSDeployType {
+			case "full-deploy":
+				err := p.gatewayClient.ServerFullDeploy()
+				if err != nil {
+					return err
+				}
+			case "no-deploy":
+				log.Debug("Not executing deploy because DNSDeployType is set to 'no-deploy'")
+			}
+		}
+	} else {
+		log.Debug("Not executing deploy because server name was not provided")
+	}
 
 	return nil
 }
@@ -359,7 +294,7 @@ func (p *BluecatProvider) zones() ([]string, error) {
 	log.Debugf("retrieving Bluecat zones for configuration: %s, view: %s", p.DNSConfiguration, p.View)
 	var zones []string
 
-	zonelist, err := p.gatewayClient.getBluecatZones(p.RootZone)
+	zonelist, err := p.gatewayClient.GetBluecatZones(p.RootZone)
 	if err != nil {
 		return nil, err
 	}
@@ -374,7 +309,7 @@ func (p *BluecatProvider) zones() ([]string, error) {
 			continue
 		}
 
-		zoneProps := splitProperties(zone.Properties)
+		zoneProps := api.SplitProperties(zone.Properties)
 
 		zones = append(zones, zoneProps["absoluteName"])
 	}
@@ -417,11 +352,11 @@ func (p *BluecatProvider) createRecords(created bluecatChangeMap) {
 			var response interface{}
 			switch ep.RecordType {
 			case endpoint.RecordTypeA:
-				response, err = p.gatewayClient.createHostRecord(zone, recordSet.obj.(*bluecatCreateHostRecordRequest))
+				err = p.gatewayClient.CreateHostRecord(zone, recordSet.obj.(*api.BluecatCreateHostRecordRequest))
 			case endpoint.RecordTypeCNAME:
-				response, err = p.gatewayClient.createCNAMERecord(zone, recordSet.obj.(*bluecatCreateCNAMERecordRequest))
+				err = p.gatewayClient.CreateCNAMERecord(zone, recordSet.obj.(*api.BluecatCreateCNAMERecordRequest))
 			case endpoint.RecordTypeTXT:
-				response, err = p.gatewayClient.createTXTRecord(zone, recordSet.obj.(*bluecatCreateTXTRecordRequest))
+				err = p.gatewayClient.CreateTXTRecord(zone, recordSet.obj.(*api.BluecatCreateTXTRecordRequest))
 			}
 			log.Debugf("Response from create: %v", response)
 			if err != nil {
@@ -471,16 +406,16 @@ func (p *BluecatProvider) deleteRecords(deleted bluecatChangeMap) {
 
 				switch ep.RecordType {
 				case endpoint.RecordTypeA:
-					for _, record := range *recordSet.res.(*[]BluecatHostRecord) {
-						err = p.gatewayClient.deleteHostRecord(record.Name, zone)
+					for _, record := range *recordSet.res.(*[]api.BluecatHostRecord) {
+						err = p.gatewayClient.DeleteHostRecord(record.Name, zone)
 					}
 				case endpoint.RecordTypeCNAME:
-					for _, record := range *recordSet.res.(*[]BluecatCNAMERecord) {
-						err = p.gatewayClient.deleteCNAMERecord(record.Name, zone)
+					for _, record := range *recordSet.res.(*[]api.BluecatCNAMERecord) {
+						err = p.gatewayClient.DeleteCNAMERecord(record.Name, zone)
 					}
 				case endpoint.RecordTypeTXT:
-					for _, record := range *recordSet.res.(*[]BluecatTXTRecord) {
-						err = p.gatewayClient.deleteTXTRecord(record.Name, zone)
+					for _, record := range *recordSet.res.(*[]api.BluecatTXTRecord) {
+						err = p.gatewayClient.DeleteTXTRecord(record.Name, zone)
 					}
 				}
 				if err != nil {
@@ -495,21 +430,22 @@ func (p *BluecatProvider) deleteRecords(deleted bluecatChangeMap) {
 	}
 }
 
-func (p *BluecatProvider) recordSet(ep *endpoint.Endpoint, getObject bool) (recordSet bluecatRecordSet, err error) {
+func (p *BluecatProvider) recordSet(ep *endpoint.Endpoint, getObject bool) (bluecatRecordSet, error) {
+	recordSet := bluecatRecordSet{}
 	switch ep.RecordType {
 	case endpoint.RecordTypeA:
-		var res []BluecatHostRecord
-		obj := bluecatCreateHostRecordRequest{
+		var res []api.BluecatHostRecord
+		obj := api.BluecatCreateHostRecordRequest{
 			AbsoluteName: ep.DNSName,
 			IP4Address:   ep.Targets[0],
 			TTL:          int(ep.RecordTTL),
 			Properties:   "",
 		}
 		if getObject {
-			var record BluecatHostRecord
-			err = p.gatewayClient.getHostRecord(ep.DNSName, &record)
+			var record api.BluecatHostRecord
+			err := p.gatewayClient.GetHostRecord(ep.DNSName, &record)
 			if err != nil {
-				return
+				return bluecatRecordSet{}, err
 			}
 			res = append(res, record)
 		}
@@ -518,18 +454,18 @@ func (p *BluecatProvider) recordSet(ep *endpoint.Endpoint, getObject bool) (reco
 			res: &res,
 		}
 	case endpoint.RecordTypeCNAME:
-		var res []BluecatCNAMERecord
-		obj := bluecatCreateCNAMERecordRequest{
+		var res []api.BluecatCNAMERecord
+		obj := api.BluecatCreateCNAMERecordRequest{
 			AbsoluteName: ep.DNSName,
 			LinkedRecord: ep.Targets[0],
 			TTL:          int(ep.RecordTTL),
 			Properties:   "",
 		}
 		if getObject {
-			var record BluecatCNAMERecord
-			err = p.gatewayClient.getCNAMERecord(ep.DNSName, &record)
+			var record api.BluecatCNAMERecord
+			err := p.gatewayClient.GetCNAMERecord(ep.DNSName, &record)
 			if err != nil {
-				return
+				return bluecatRecordSet{}, err
 			}
 			res = append(res, record)
 		}
@@ -538,18 +474,18 @@ func (p *BluecatProvider) recordSet(ep *endpoint.Endpoint, getObject bool) (reco
 			res: &res,
 		}
 	case endpoint.RecordTypeTXT:
-		var res []BluecatTXTRecord
+		var res []api.BluecatTXTRecord
 		// TODO: Allow setting TTL
 		// This is not implemented in the Bluecat Gateway
-		obj := bluecatCreateTXTRecordRequest{
+		obj := api.BluecatCreateTXTRecordRequest{
 			AbsoluteName: ep.DNSName,
 			Text:         ep.Targets[0],
 		}
 		if getObject {
-			var record BluecatTXTRecord
-			err = p.gatewayClient.getTXTRecord(ep.DNSName, &record)
+			var record api.BluecatTXTRecord
+			err := p.gatewayClient.GetTXTRecord(ep.DNSName, &record)
 			if err != nil {
-				return
+				return bluecatRecordSet{}, err
 			}
 			res = append(res, record)
 		}
@@ -558,479 +494,10 @@ func (p *BluecatProvider) recordSet(ep *endpoint.Endpoint, getObject bool) (reco
 			res: &res,
 		}
 	}
-	return
+	return recordSet, nil
 }
 
-// getBluecatGatewayToken retrieves a Bluecat Gateway API token.
-func getBluecatGatewayToken(cfg bluecatConfig) (string, http.Cookie, error) {
-	var username string
-	if cfg.GatewayUsername != "" {
-		username = cfg.GatewayUsername
-	}
-	if v, ok := os.LookupEnv("BLUECAT_USERNAME"); ok {
-		username = v
-	}
-
-	var password string
-	if cfg.GatewayPassword != "" {
-		password = cfg.GatewayPassword
-	}
-	if v, ok := os.LookupEnv("BLUECAT_PASSWORD"); ok {
-		password = v
-	}
-
-	body, err := json.Marshal(map[string]string{
-		"username": username,
-		"password": password,
-	})
-	if err != nil {
-		return "", http.Cookie{}, errors.Wrap(err, "could not unmarshal credentials for bluecat gateway config")
-	}
-
-	c := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: cfg.SkipTLSVerify},
-		}}
-
-	resp, err := c.Post(cfg.GatewayHost+"/rest_login", "application/json", bytes.NewBuffer(body))
-	if err != nil {
-		return "", http.Cookie{}, errors.Wrap(err, "error obtaining API token from bluecat gateway")
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		details, _ := ioutil.ReadAll(resp.Body)
-		return "", http.Cookie{}, errors.Errorf("got HTTP response code %v, detailed message: %v", resp.StatusCode, string(details))
-	}
-
-	res, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", http.Cookie{}, errors.Wrap(err, "error reading get_token response from bluecat gateway")
-	}
-
-	resJSON := map[string]string{}
-	err = json.Unmarshal(res, &resJSON)
-	if err != nil {
-		return "", http.Cookie{}, errors.Wrap(err, "error unmarshaling json response (auth) from bluecat gateway")
-	}
-
-	// Example response: {"access_token": "BAMAuthToken: abc123"}
-	// We only care about the actual token string - i.e. abc123
-	// The gateway also creates a cookie as part of the response. This seems to be the actual auth mechanism, at least
-	// for now.
-	return strings.Split(resJSON["access_token"], " ")[1], *resp.Cookies()[0], nil
-}
-
-func (c GatewayClientConfig) getBluecatZones(zoneName string) ([]BluecatZone, error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-	zonePath := expandZone(zoneName)
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "error building http request")
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, errors.Wrapf(err, "error retrieving zone(s) from gateway: %v, %v", url, zoneName)
-	}
-
-	defer resp.Body.Close()
-
-	zones := []BluecatZone{}
-	json.NewDecoder(resp.Body).Decode(&zones)
-
-	// Bluecat Gateway only returns subzones one level deeper than the provided zone
-	// so this recursion is needed to traverse subzones until none are returned
-	for _, zone := range zones {
-		zoneProps := splitProperties(zone.Properties)
-		subZones, err := c.getBluecatZones(zoneProps["absoluteName"])
-		if err != nil {
-			return nil, errors.Wrapf(err, "error retrieving subzones from gateway: %v", zoneName)
-		}
-		zones = append(zones, subZones...)
-	}
-
-	return zones, nil
-}
-
-func (c GatewayClientConfig) getHostRecords(zone string, records *[]BluecatHostRecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "host_records/"
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "error building http request")
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", zone)
-	}
-
-	defer resp.Body.Close()
-
-	json.NewDecoder(resp.Body).Decode(records)
-	log.Debugf("Get Host Records Response: %v", records)
-
-	return nil
-}
-
-func (c GatewayClientConfig) getCNAMERecords(zone string, records *[]BluecatCNAMERecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "cname_records/"
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "error building http request")
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", zone)
-	}
-
-	defer resp.Body.Close()
-
-	json.NewDecoder(resp.Body).Decode(records)
-	log.Debugf("Get CName Records Response: %v", records)
-
-	return nil
-}
-
-func (c GatewayClientConfig) getTXTRecords(zone string, records *[]BluecatTXTRecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "text_records/"
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "error building http request")
-	}
-	log.Debugf("Request: %v", req)
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", zone)
-	}
-	log.Debugf("Get Txt Records response: %v", resp)
-
-	defer resp.Body.Close()
-	json.NewDecoder(resp.Body).Decode(records)
-	log.Debugf("Get TXT Records Body: %v", records)
-
-	return nil
-}
-
-func (c GatewayClientConfig) getHostRecord(name string, record *BluecatHostRecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"host_records/" + name + "/"
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrapf(err, "error building http request: %v", name)
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", name)
-	}
-
-	defer resp.Body.Close()
-
-	json.NewDecoder(resp.Body).Decode(record)
-	log.Debugf("Get Host Record Response: %v", record)
-	return nil
-}
-
-func (c GatewayClientConfig) getCNAMERecord(name string, record *BluecatCNAMERecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"cname_records/" + name + "/"
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrapf(err, "error building http request: %v", name)
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", name)
-	}
-
-	defer resp.Body.Close()
-
-	json.NewDecoder(resp.Body).Decode(record)
-	log.Debugf("Get CName Record Response: %v", record)
-	return nil
-}
-
-func (c GatewayClientConfig) getTXTRecord(name string, record *BluecatTXTRecord) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"text_records/" + name + "/"
-
-	req, err := c.buildHTTPRequest("GET", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "error building http request")
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error retrieving record(s) from gateway: %v", name)
-	}
-
-	defer resp.Body.Close()
-	json.NewDecoder(resp.Body).Decode(record)
-	log.Debugf("Get TXT Record Response: %v", record)
-
-	return nil
-}
-
-func (c GatewayClientConfig) createHostRecord(zone string, req *bluecatCreateHostRecordRequest) (res interface{}, err error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "host_records/"
-	body, _ := json.Marshal(req)
-	hreq, err := c.buildHTTPRequest("POST", url, bytes.NewBuffer(body))
-	if err != nil {
-		return nil, errors.Wrap(err, "error building http request")
-	}
-	hreq.Header.Add("Content-Type", "application/json")
-	res, err = client.Do(hreq)
-
-	return
-}
-
-func (c GatewayClientConfig) createCNAMERecord(zone string, req *bluecatCreateCNAMERecordRequest) (res interface{}, err error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "cname_records/"
-	body, _ := json.Marshal(req)
-
-	hreq, err := c.buildHTTPRequest("POST", url, bytes.NewBuffer(body))
-	if err != nil {
-		return nil, errors.Wrap(err, "error building http request")
-	}
-
-	hreq.Header.Add("Content-Type", "application/json")
-	res, err = client.Do(hreq)
-
-	return
-}
-
-func (c GatewayClientConfig) createTXTRecord(zone string, req *bluecatCreateTXTRecordRequest) (interface{}, error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	zonePath := expandZone(zone)
-	// Remove the trailing 'zones/'
-	zonePath = strings.TrimSuffix(zonePath, "zones/")
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration + "/views/" + c.View + "/" + zonePath + "text_records/"
-	body, _ := json.Marshal(req)
-	hreq, err := c.buildHTTPRequest("POST", url, bytes.NewBuffer(body))
-	if err != nil {
-		return nil, err
-	}
-
-	hreq.Header.Add("Content-Type", "application/json")
-	res, err := client.Do(hreq)
-
-	return res, err
-}
-
-func (c GatewayClientConfig) deleteHostRecord(name string, zone string) (err error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"host_records/" + name + "." + zone + "/"
-	req, err := c.buildHTTPRequest("DELETE", url, nil)
-	if err != nil {
-		return errors.Wrapf(err, "error building http request: %v", name)
-	}
-
-	_, err = client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error deleting record(s) from gateway: %v", name)
-	}
-
-	return nil
-}
-
-func (c GatewayClientConfig) deleteCNAMERecord(name string, zone string) (err error) {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"cname_records/" + name + "." + zone + "/"
-	req, err := c.buildHTTPRequest("DELETE", url, nil)
-	if err != nil {
-		return errors.Wrapf(err, "error building http request: %v", name)
-	}
-
-	_, err = client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error deleting record(s) from gateway: %v", name)
-	}
-
-	return nil
-}
-
-func (c GatewayClientConfig) deleteTXTRecord(name string, zone string) error {
-	transportCfg := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipTLSVerify},
-	}
-	client := &http.Client{
-		Transport: transportCfg,
-	}
-
-	url := c.Host + "/api/v1/configurations/" + c.DNSConfiguration +
-		"/views/" + c.View + "/" +
-		"text_records/" + name + "." + zone + "/"
-
-	req, err := c.buildHTTPRequest("DELETE", url, nil)
-	if err != nil {
-		return errors.Wrap(err, "error building http request")
-	}
-
-	_, err = client.Do(req)
-	if err != nil {
-		return errors.Wrapf(err, "error deleting record(s) from gateway: %v", name)
-	}
-
-	return nil
-}
-
-//buildHTTPRequest builds a standard http Request and adds authentication headers required by Bluecat Gateway
-func (c GatewayClientConfig) buildHTTPRequest(method, url string, body io.Reader) (*http.Request, error) {
-	req, err := http.NewRequest(method, url, body)
-	req.Header.Add("Accept", "application/json")
-	req.Header.Add("Authorization", "Basic "+c.Token)
-	req.AddCookie(&c.Cookie)
-	return req, err
-}
-
-//splitProperties is a helper function to break a '|' separated string into key/value pairs
-// i.e. "foo=bar|baz=mop"
-func splitProperties(props string) map[string]string {
-	propMap := make(map[string]string)
-	// remove trailing | character before we split
-	props = strings.TrimSuffix(props, "|")
-
-	splits := strings.Split(props, "|")
-	for _, pair := range splits {
-		items := strings.Split(pair, "=")
-		propMap[items[0]] = items[1]
-	}
-
-	return propMap
-}
-
-//expandZone takes an absolute domain name such as 'example.com' and returns a zone hierarchy used by Bluecat Gateway,
-//such as '/zones/com/zones/example/zones/'
-func expandZone(zone string) string {
-	ze := "zones/"
-	parts := strings.Split(zone, ".")
-	if len(parts) > 1 {
-		last := len(parts) - 1
-		for i := range parts {
-			ze = ze + parts[last-i] + "/zones/"
-		}
-	} else {
-		ze = ze + zone + "/zones/"
-	}
-	return ze
-}
-
-//extractOwnerFromTXTRecord takes a single text property string and returns the owner after parsing theowner string.
+// extractOwnerFromTXTRecord takes a single text property string and returns the owner after parsing the owner string.
 func extractOwnerfromTXTRecord(propString string) (string, error) {
 	if len(propString) == 0 {
 		return "", errors.Errorf("External-DNS Owner not found")
