@@ -20,11 +20,12 @@ import (
 	"context"
 	"encoding/base64"
 	"fmt"
+	"net/http"
 	"regexp"
 	"strings"
 	"testing"
 
-	ibclient "github.com/infobloxopen/infoblox-go-client"
+	ibclient "github.com/infobloxopen/infoblox-go-client/v2"
 	"github.com/miekg/dns"
 	"github.com/stretchr/testify/assert"
 
@@ -113,7 +114,7 @@ func (client *mockIBConnector) CreateObject(obj ibclient.IBObject) (ref string, 
 	return ref, nil
 }
 
-func (client *mockIBConnector) GetObject(obj ibclient.IBObject, ref string, res interface{}) (err error) {
+func (client *mockIBConnector) GetObject(obj ibclient.IBObject, ref string, queryParams *ibclient.QueryParams, res interface{}) (err error) {
 	switch obj.ObjectType() {
 	case "record:a":
 		var result []ibclient.RecordA
@@ -208,12 +209,9 @@ func (client *mockIBConnector) DeleteObject(ref string) (refRes string, err erro
 	switch result[1] {
 	case "record:a":
 		var records []ibclient.RecordA
-		obj := ibclient.NewRecordA(
-			ibclient.RecordA{
-				Name: result[2],
-			},
-		)
-		client.GetObject(obj, ref, &records)
+		obj := ibclient.NewEmptyRecordA()
+		obj.Name = result[2]
+		client.GetObject(obj, ref, nil, &records)
 		for _, record := range records {
 			client.deletedEndpoints = append(
 				client.deletedEndpoints,
@@ -226,12 +224,9 @@ func (client *mockIBConnector) DeleteObject(ref string) (refRes string, err erro
 		}
 	case "record:cname":
 		var records []ibclient.RecordCNAME
-		obj := ibclient.NewRecordCNAME(
-			ibclient.RecordCNAME{
-				Name: result[2],
-			},
-		)
-		client.GetObject(obj, ref, &records)
+		obj := ibclient.NewEmptyRecordCNAME()
+		obj.Name = result[2]
+		client.GetObject(obj, ref, nil, &records)
 		for _, record := range records {
 			client.deletedEndpoints = append(
 				client.deletedEndpoints,
@@ -244,12 +239,9 @@ func (client *mockIBConnector) DeleteObject(ref string) (refRes string, err erro
 		}
 	case "record:host":
 		var records []ibclient.HostRecord
-		obj := ibclient.NewHostRecord(
-			ibclient.HostRecord{
-				Name: result[2],
-			},
-		)
-		client.GetObject(obj, ref, &records)
+		obj := ibclient.NewEmptyHostRecord()
+		obj.Name = result[2]
+		client.GetObject(obj, ref, nil, &records)
 		for _, record := range records {
 			client.deletedEndpoints = append(
 				client.deletedEndpoints,
@@ -267,7 +259,7 @@ func (client *mockIBConnector) DeleteObject(ref string) (refRes string, err erro
 				Name: result[2],
 			},
 		)
-		client.GetObject(obj, ref, &records)
+		client.GetObject(obj, ref, nil, &records)
 		for _, record := range records {
 			client.deletedEndpoints = append(
 				client.deletedEndpoints,
@@ -280,12 +272,9 @@ func (client *mockIBConnector) DeleteObject(ref string) (refRes string, err erro
 		}
 	case "record:ptr":
 		var records []ibclient.RecordPTR
-		obj := ibclient.NewRecordPTR(
-			ibclient.RecordPTR{
-				Name: result[2],
-			},
-		)
-		client.GetObject(obj, ref, &records)
+		obj := ibclient.NewEmptyRecordPTR()
+		obj.Name = result[2]
+		client.GetObject(obj, ref, nil, &records)
 		for _, record := range records {
 			client.deletedEndpoints = append(
 				client.deletedEndpoints,
@@ -354,21 +343,17 @@ func createMockInfobloxObject(name, recordType, value string) ibclient.IBObject 
 	ref := fmt.Sprintf("record:%s/%s:%s/default", strings.ToLower(recordType), base64.StdEncoding.EncodeToString([]byte(name)), name)
 	switch recordType {
 	case endpoint.RecordTypeA:
-		return ibclient.NewRecordA(
-			ibclient.RecordA{
-				Ref:      ref,
-				Name:     name,
-				Ipv4Addr: value,
-			},
-		)
+		obj := ibclient.NewEmptyRecordA()
+		obj.Name = name
+		obj.Ref = ref
+		obj.Ipv4Addr = value
+		return obj
 	case endpoint.RecordTypeCNAME:
-		return ibclient.NewRecordCNAME(
-			ibclient.RecordCNAME{
-				Ref:       ref,
-				Name:      name,
-				Canonical: value,
-			},
-		)
+		obj := ibclient.NewEmptyRecordCNAME()
+		obj.Name = name
+		obj.Ref = ref
+		obj.Canonical = value
+		return obj
 	case endpoint.RecordTypeTXT:
 		return ibclient.NewRecordTXT(
 			ibclient.RecordTXT{
@@ -378,31 +363,28 @@ func createMockInfobloxObject(name, recordType, value string) ibclient.IBObject 
 			},
 		)
 	case "HOST":
-		return ibclient.NewHostRecord(
-			ibclient.HostRecord{
-				Ref:  ref,
-				Name: name,
-				Ipv4Addrs: []ibclient.HostRecordIpv4Addr{
-					{
-						Ipv4Addr: value,
-					},
-				},
-			},
-		)
-	case endpoint.RecordTypePTR:
-		return ibclient.NewRecordPTR(
-			ibclient.RecordPTR{
-				Ref:      ref,
-				PtrdName: name,
+		obj := ibclient.NewEmptyHostRecord()
+		obj.Name = name
+		obj.Ref = ref
+		obj.Ipv4Addrs = []ibclient.HostRecordIpv4Addr{
+			{
 				Ipv4Addr: value,
 			},
-		)
+		}
+		return obj
+	case endpoint.RecordTypePTR:
+		obj := ibclient.NewEmptyRecordPTR()
+		obj.PtrdName = name
+		obj.Ref = ref
+		obj.Ipv4Addr = value
+		return obj
 	}
+
 	return nil
 }
 
-func newInfobloxProvider(domainFilter endpoint.DomainFilter, zoneIDFilter provider.ZoneIDFilter, dryRun bool, createPTR bool, client ibclient.IBConnector) *InfobloxProvider {
-	return &InfobloxProvider{
+func newInfobloxProvider(domainFilter endpoint.DomainFilter, zoneIDFilter provider.ZoneIDFilter, dryRun bool, createPTR bool, client ibclient.IBConnector) *ProviderConfig {
+	return &ProviderConfig{
 		client:       client,
 		domainFilter: domainFilter,
 		zoneIDFilter: zoneIDFilter,
@@ -435,8 +417,8 @@ func TestInfobloxRecords(t *testing.T) {
 		},
 	}
 
-	provider := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
-	actual, err := provider.Records(context.Background())
+	providerCfg := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
+	actual, err := providerCfg.Records(context.Background())
 
 	if err != nil {
 		t.Fatal(err)
@@ -472,12 +454,12 @@ func TestInfobloxAdjustEndpoints(t *testing.T) {
 		},
 	}
 
-	provider := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com"}), provider.NewZoneIDFilter([]string{""}), true, true, &client)
-	actual, err := provider.Records(context.Background())
+	providerCfg := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com"}), provider.NewZoneIDFilter([]string{""}), true, true, &client)
+	actual, err := providerCfg.Records(context.Background())
 	if err != nil {
 		t.Fatal(err)
 	}
-	provider.AdjustEndpoints(actual)
+	providerCfg.AdjustEndpoints(actual)
 
 	expected := []*endpoint.Endpoint{
 		endpoint.NewEndpoint("example.com", endpoint.RecordTypeA, "123.123.123.122").WithProviderSpecific(providerSpecificInfobloxPtrRecord, "true"),
@@ -501,8 +483,8 @@ func TestInfobloxRecordsReverse(t *testing.T) {
 		},
 	}
 
-	provider := newInfobloxProvider(endpoint.NewDomainFilter([]string{"10.0.0.0/24"}), provider.NewZoneIDFilter([]string{""}), true, true, &client)
-	actual, err := provider.Records(context.Background())
+	providerCfg := newInfobloxProvider(endpoint.NewDomainFilter([]string{"10.0.0.0/24"}), provider.NewZoneIDFilter([]string{""}), true, true, &client)
+	actual, err := providerCfg.Records(context.Background())
 
 	if err != nil {
 		t.Fatal(err)
@@ -606,7 +588,7 @@ func testInfobloxApplyChangesInternal(t *testing.T, dryRun, createPTR bool, clie
 		createMockInfobloxObject("oldcname.example.com", endpoint.RecordTypeCNAME, "other.com"),
 	}
 
-	provider := newInfobloxProvider(
+	providerCfg := newInfobloxProvider(
 		endpoint.NewDomainFilter([]string{""}),
 		provider.NewZoneIDFilter([]string{""}),
 		dryRun,
@@ -658,7 +640,7 @@ func testInfobloxApplyChangesInternal(t *testing.T, dryRun, createPTR bool, clie
 		Delete:    deleteRecords,
 	}
 
-	if err := provider.ApplyChanges(context.Background(), changes); err != nil {
+	if err := providerCfg.ApplyChanges(context.Background(), changes); err != nil {
 		t.Fatal(err)
 	}
 }
@@ -674,18 +656,18 @@ func TestInfobloxZones(t *testing.T) {
 		mockInfobloxObjects: &[]ibclient.IBObject{},
 	}
 
-	provider := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com", "1.2.3.0/24"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
-	zones, _ := provider.zones()
+	providerCfg := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com", "1.2.3.0/24"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
+	zones, _ := providerCfg.zones()
 	var emptyZoneAuth *ibclient.ZoneAuth
-	assert.Equal(t, provider.findZone(zones, "example.com").Fqdn, "example.com")
-	assert.Equal(t, provider.findZone(zones, "nomatch-example.com"), emptyZoneAuth)
-	assert.Equal(t, provider.findZone(zones, "nginx.example.com").Fqdn, "example.com")
-	assert.Equal(t, provider.findZone(zones, "lvl1-1.example.com").Fqdn, "lvl1-1.example.com")
-	assert.Equal(t, provider.findZone(zones, "lvl1-2.example.com").Fqdn, "example.com")
-	assert.Equal(t, provider.findZone(zones, "lvl2-1.lvl1-1.example.com").Fqdn, "lvl2-1.lvl1-1.example.com")
-	assert.Equal(t, provider.findZone(zones, "lvl2-2.lvl1-1.example.com").Fqdn, "lvl1-1.example.com")
-	assert.Equal(t, provider.findZone(zones, "lvl2-2.lvl1-2.example.com").Fqdn, "example.com")
-	assert.Equal(t, provider.findZone(zones, "1.2.3.0/24").Fqdn, "1.2.3.0/24")
+	assert.Equal(t, providerCfg.findZone(zones, "example.com").Fqdn, "example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "nomatch-example.com"), emptyZoneAuth)
+	assert.Equal(t, providerCfg.findZone(zones, "nginx.example.com").Fqdn, "example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "lvl1-1.example.com").Fqdn, "lvl1-1.example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "lvl1-2.example.com").Fqdn, "example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "lvl2-1.lvl1-1.example.com").Fqdn, "lvl2-1.lvl1-1.example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "lvl2-2.lvl1-1.example.com").Fqdn, "lvl1-1.example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "lvl2-2.lvl1-2.example.com").Fqdn, "example.com")
+	assert.Equal(t, providerCfg.findZone(zones, "1.2.3.0/24").Fqdn, "1.2.3.0/24")
 }
 
 func TestInfobloxReverseZones(t *testing.T) {
@@ -698,58 +680,98 @@ func TestInfobloxReverseZones(t *testing.T) {
 		mockInfobloxObjects: &[]ibclient.IBObject{},
 	}
 
-	provider := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com", "1.2.3.0/24", "10.0.0.0/8"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
-	zones, _ := provider.zones()
+	providerCfg := newInfobloxProvider(endpoint.NewDomainFilter([]string{"example.com", "1.2.3.0/24", "10.0.0.0/8"}), provider.NewZoneIDFilter([]string{""}), true, false, &client)
+	zones, _ := providerCfg.zones()
 	var emptyZoneAuth *ibclient.ZoneAuth
-	assert.Equal(t, provider.findReverseZone(zones, "nomatch-example.com"), emptyZoneAuth)
-	assert.Equal(t, provider.findReverseZone(zones, "192.168.0.1"), emptyZoneAuth)
-	assert.Equal(t, provider.findReverseZone(zones, "1.2.3.4").Fqdn, "1.2.3.0/24")
-	assert.Equal(t, provider.findReverseZone(zones, "10.28.29.30").Fqdn, "10.0.0.0/8")
+	assert.Equal(t, providerCfg.findReverseZone(zones, "nomatch-example.com"), emptyZoneAuth)
+	assert.Equal(t, providerCfg.findReverseZone(zones, "192.168.0.1"), emptyZoneAuth)
+	assert.Equal(t, providerCfg.findReverseZone(zones, "1.2.3.4").Fqdn, "1.2.3.0/24")
+	assert.Equal(t, providerCfg.findReverseZone(zones, "10.28.29.30").Fqdn, "10.0.0.0/8")
 }
 
 func TestExtendedRequestFDQDRegExBuilder(t *testing.T) {
-	hostConfig := ibclient.HostConfig{
-		Host:     "localhost",
-		Port:     "8080",
+	hostCfg := ibclient.HostConfig{
+		Host:    "localhost",
+		Port:    "8080",
+		Version: "2.3.1",
+	}
+
+	authCfg := ibclient.AuthConfig{
 		Username: "user",
 		Password: "abcd",
-		Version:  "2.3.1",
 	}
 
 	requestBuilder := NewExtendedRequestBuilder(0, "^staging.*test.com$")
-	requestBuilder.Init(hostConfig)
+	requestBuilder.Init(hostCfg, authCfg)
 
 	obj := ibclient.NewZoneAuth(ibclient.ZoneAuth{})
 
-	req, _ := requestBuilder.BuildRequest(ibclient.GET, obj, "", ibclient.QueryParams{})
+	req, _ := requestBuilder.BuildRequest(ibclient.GET, obj, "", &ibclient.QueryParams{})
 
 	assert.True(t, req.URL.Query().Get("fqdn~") == "^staging.*test.com$")
 
-	req, _ = requestBuilder.BuildRequest(ibclient.CREATE, obj, "", ibclient.QueryParams{})
+	req, _ = requestBuilder.BuildRequest(ibclient.CREATE, obj, "", &ibclient.QueryParams{})
 
 	assert.True(t, req.URL.Query().Get("fqdn~") == "")
 }
+
 func TestExtendedRequestMaxResultsBuilder(t *testing.T) {
-	hostConfig := ibclient.HostConfig{
-		Host:     "localhost",
-		Port:     "8080",
+	hostCfg := ibclient.HostConfig{
+		Host:    "localhost",
+		Port:    "8080",
+		Version: "2.3.1",
+	}
+
+	authCfg := ibclient.AuthConfig{
 		Username: "user",
 		Password: "abcd",
-		Version:  "2.3.1",
 	}
 
 	requestBuilder := NewExtendedRequestBuilder(54321, "")
-	requestBuilder.Init(hostConfig)
+	requestBuilder.Init(hostCfg, authCfg)
 
-	obj := ibclient.NewRecordCNAME(ibclient.RecordCNAME{Zone: "foo.bar.com"})
+	obj := ibclient.NewEmptyRecordCNAME()
+	obj.Zone = "foo.bar.com"
 
-	req, _ := requestBuilder.BuildRequest(ibclient.GET, obj, "", ibclient.QueryParams{})
+	req, _ := requestBuilder.BuildRequest(ibclient.GET, obj, "", &ibclient.QueryParams{})
 
 	assert.True(t, req.URL.Query().Get("_max_results") == "54321")
 
-	req, _ = requestBuilder.BuildRequest(ibclient.CREATE, obj, "", ibclient.QueryParams{})
+	req, _ = requestBuilder.BuildRequest(ibclient.CREATE, obj, "", &ibclient.QueryParams{})
 
 	assert.True(t, req.URL.Query().Get("_max_results") == "")
+}
+
+func TestGetObject(t *testing.T) {
+	hostCfg := ibclient.HostConfig{}
+	authCfg := ibclient.AuthConfig{}
+	transportConfig := ibclient.TransportConfig{}
+	requestBuilder := NewExtendedRequestBuilder(1000, "mysite.com")
+	requestor := mockRequestor{}
+	client, _ := ibclient.NewConnector(hostCfg, authCfg, transportConfig, requestBuilder, &requestor)
+
+	providerConfig := newInfobloxProvider(endpoint.NewDomainFilter([]string{"mysite.com"}), provider.NewZoneIDFilter([]string{""}), true, true, client)
+
+	providerConfig.deleteRecords(infobloxChangeMap{
+		"myzone.com": []*endpoint.Endpoint{
+			endpoint.NewEndpoint("deletethisrecord.com", endpoint.RecordTypeA, "1.2.3.4"),
+		},
+	})
+
+	requestQuery := requestor.request.URL.Query()
+	assert.True(t, requestQuery.Has("name"), "Expected the request to filter objects by name")
+}
+
+// Mock requestor that doesn't send request
+type mockRequestor struct {
+	request *http.Request
+}
+
+func (r *mockRequestor) Init(ibclient.AuthConfig, ibclient.TransportConfig) {}
+func (r *mockRequestor) SendRequest(req *http.Request) (res []byte, err error) {
+	res = []byte("[{}]")
+	r.request = req
+	return
 }
 
 func validateEndpoints(t *testing.T, endpoints []*endpoint.Endpoint, expected []*endpoint.Endpoint) {
