@@ -27,6 +27,7 @@ import (
 
 	routev1 "github.com/openshift/api/route/v1"
 	fake "github.com/openshift/client-go/route/clientset/versioned/fake"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"sigs.k8s.io/external-dns/endpoint"
@@ -43,6 +44,7 @@ func (suite *OCPRouteSuite) SetupTest() {
 	var err error
 
 	suite.sc, err = NewOcpRouteSource(
+		context.TODO(),
 		fakeClient,
 		"",
 		"",
@@ -50,6 +52,7 @@ func (suite *OCPRouteSuite) SetupTest() {
 		false,
 		false,
 		labels.Everything(),
+		"",
 	)
 
 	suite.routeWithTargets = &routev1.Route{
@@ -140,6 +143,7 @@ func testOcpRouteSourceNewOcpRouteSource(t *testing.T) {
 			t.Parallel()
 
 			_, err := NewOcpRouteSource(
+				context.TODO(),
 				fake.NewSimpleClientset(),
 				"",
 				ti.annotationFilter,
@@ -147,6 +151,7 @@ func testOcpRouteSourceNewOcpRouteSource(t *testing.T) {
 				false,
 				false,
 				labelSelector,
+				"",
 			)
 
 			if ti.expectError {
@@ -160,38 +165,32 @@ func testOcpRouteSourceNewOcpRouteSource(t *testing.T) {
 
 // testOcpRouteSourceEndpoints tests that various OCP routes generate the correct endpoints.
 func testOcpRouteSourceEndpoints(t *testing.T) {
-	t.Parallel()
-
 	for _, tc := range []struct {
-		title                    string
-		targetNamespace          string
-		annotationFilter         string
-		fqdnTemplate             string
-		ignoreHostnameAnnotation bool
-		ocpRoute                 *routev1.Route
-		expected                 []*endpoint.Endpoint
-		expectError              bool
-		labelFilter              string
+		title         string
+		ocpRoute      *routev1.Route
+		expected      []*endpoint.Endpoint
+		expectError   bool
+		labelFilter   string
+		ocpRouterName string
 	}{
 		{
-			title:                    "route with basic hostname and route status target",
-			targetNamespace:          "",
-			annotationFilter:         "",
-			fqdnTemplate:             "",
-			ignoreHostnameAnnotation: false,
+			title: "route with basic hostname and route status target",
 			ocpRoute: &routev1.Route{
-				Spec: routev1.RouteSpec{
-					Host: "my-domain.com",
-				},
 				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "default",
-					Name:        "route-with-target",
-					Annotations: map[string]string{},
+					Namespace: "default",
+					Name:      "route-with-target",
 				},
 				Status: routev1.RouteStatus{
 					Ingress: []routev1.RouteIngress{
 						{
+							Host:                    "my-domain.com",
 							RouterCanonicalHostname: "apps.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
 						},
 					},
 				},
@@ -205,14 +204,203 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 					},
 				},
 			},
-			expectError: false,
 		},
 		{
-			title:                    "route with incorrect externalDNS controller annotation",
-			targetNamespace:          "",
-			annotationFilter:         "",
-			fqdnTemplate:             "",
-			ignoreHostnameAnnotation: false,
+			title: "route with basic hostname, route status target and ocpRouterName defined",
+			ocpRoute: &routev1.Route{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "route-with-target",
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+					},
+				},
+			},
+			ocpRouterName: "default",
+			expected: []*endpoint.Endpoint{
+				{
+					DNSName: "my-domain.com",
+					Targets: []string{
+						"router-default.my-domain.com",
+					},
+				},
+			},
+		},
+		{
+			title: "route with basic hostname, route status target, one ocpRouterName and two router canonical names",
+			ocpRoute: &routev1.Route{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "route-with-target",
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "test",
+							RouterCanonicalHostname: "router-test.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+					},
+				},
+			},
+			ocpRouterName: "default",
+			expected: []*endpoint.Endpoint{
+				{
+					DNSName: "my-domain.com",
+					Targets: []string{
+						"router-default.my-domain.com",
+					},
+				},
+			},
+		},
+		{
+			title: "route not admitted by the given router",
+			ocpRoute: &routev1.Route{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "route-with-target",
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "test",
+							RouterCanonicalHostname: "router-test.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionFalse,
+								},
+							},
+						},
+					},
+				},
+			},
+			ocpRouterName: "test",
+			expected:      []*endpoint.Endpoint{},
+		},
+		{
+			title: "route not admitted by any router",
+			ocpRoute: &routev1.Route{
+				Spec: routev1.RouteSpec{
+					Host: "my-domain.com",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "route-with-target",
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionFalse,
+								},
+							},
+						},
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "test",
+							RouterCanonicalHostname: "router-test.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionFalse,
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: []*endpoint.Endpoint{},
+		},
+		{
+			title: "route admitted by first appropriate router",
+			ocpRoute: &routev1.Route{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "route-with-target",
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionFalse,
+								},
+							},
+						},
+						{
+							Host:                    "my-domain.com",
+							RouterName:              "test",
+							RouterCanonicalHostname: "router-test.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: []*endpoint.Endpoint{
+				{
+					DNSName: "my-domain.com",
+					Targets: []string{"router-test.my-domain.com"},
+				},
+			},
+		},
+		{
+			title: "route with incorrect externalDNS controller annotation",
 			ocpRoute: &routev1.Route{
 				ObjectMeta: metav1.ObjectMeta{
 					Namespace: "default",
@@ -222,24 +410,31 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 					},
 				},
 			},
-			expected:    []*endpoint.Endpoint{},
-			expectError: false,
+			expected: []*endpoint.Endpoint{},
 		},
 		{
-			title:                    "route with basic hostname and annotation target",
-			targetNamespace:          "",
-			annotationFilter:         "",
-			fqdnTemplate:             "",
-			ignoreHostnameAnnotation: false,
+			title: "route with basic hostname and annotation target",
 			ocpRoute: &routev1.Route{
-				Spec: routev1.RouteSpec{
-					Host: "my-annotation-domain.com",
-				},
 				ObjectMeta: metav1.ObjectMeta{
 					Namespace: "default",
 					Name:      "route-with-annotation-target",
 					Annotations: map[string]string{
 						"external-dns.alpha.kubernetes.io/target": "my.site.foo.com",
+					},
+				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-annotation-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
 					},
 				},
 			},
@@ -252,17 +447,11 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 					},
 				},
 			},
-			expectError: false,
 		},
 		{
-			title:                    "route with matching labels",
-			labelFilter:              "app=web-external",
-			ignoreHostnameAnnotation: false,
+			title:       "route with matching labels",
+			labelFilter: "app=web-external",
 			ocpRoute: &routev1.Route{
-
-				Spec: routev1.RouteSpec{
-					Host: "my-annotation-domain.com",
-				},
 				ObjectMeta: metav1.ObjectMeta{
 					Namespace: "default",
 					Name:      "route-with-matching-labels",
@@ -274,6 +463,21 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 						"name": "service-frontend",
 					},
 				},
+				Status: routev1.RouteStatus{
+					Ingress: []routev1.RouteIngress{
+						{
+							Host:                    "my-annotation-domain.com",
+							RouterName:              "default",
+							RouterCanonicalHostname: "router-default.my-domain.com",
+							Conditions: []routev1.RouteIngressCondition{
+								{
+									Type:   routev1.RouteAdmitted,
+									Status: corev1.ConditionTrue,
+								},
+							},
+						},
+					},
+				},
 			},
 			expected: []*endpoint.Endpoint{
 				{
@@ -284,12 +488,10 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 					},
 				},
 			},
-			expectError: false,
 		},
 		{
-			title:                    "route without matching labels",
-			labelFilter:              "app=web-external",
-			ignoreHostnameAnnotation: false,
+			title:       "route without matching labels",
+			labelFilter: "app=web-external",
 			ocpRoute: &routev1.Route{
 
 				Spec: routev1.RouteSpec{
@@ -307,17 +509,14 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 					},
 				},
 			},
-			expected:    []*endpoint.Endpoint{},
-			expectError: false,
+			expected: []*endpoint.Endpoint{},
 		},
 	} {
 		tc := tc
 		t.Run(tc.title, func(t *testing.T) {
 			t.Parallel()
-
 			// Create a Kubernetes testing client
 			fakeClient := fake.NewSimpleClientset()
-
 			_, err := fakeClient.RouteV1().Routes(tc.ocpRoute.Namespace).Create(context.Background(), tc.ocpRoute, metav1.CreateOptions{})
 			require.NoError(t, err)
 
@@ -325,6 +524,7 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 			require.NoError(t, err)
 
 			source, err := NewOcpRouteSource(
+				context.TODO(),
 				fakeClient,
 				"",
 				"",
@@ -332,6 +532,7 @@ func testOcpRouteSourceEndpoints(t *testing.T) {
 				false,
 				false,
 				labelSelector,
+				tc.ocpRouterName,
 			)
 			require.NoError(t, err)
 
