@@ -17,6 +17,7 @@ limitations under the License.
 package source
 
 import (
+	"context"
 	"errors"
 	"testing"
 
@@ -32,11 +33,13 @@ import (
 	fakeDynamic "k8s.io/client-go/dynamic/fake"
 	"k8s.io/client-go/kubernetes"
 	fakeKube "k8s.io/client-go/kubernetes/fake"
+	gateway "sigs.k8s.io/gateway-api/pkg/client/clientset/gateway/versioned"
 )
 
 type MockClientGenerator struct {
 	mock.Mock
 	kubeClient              kubernetes.Interface
+	gatewayClient           gateway.Interface
 	istioClient             istioclient.Interface
 	cloudFoundryClient      *cfclient.Client
 	dynamicKubernetesClient dynamic.Interface
@@ -50,6 +53,15 @@ func (m *MockClientGenerator) KubeClient() (kubernetes.Interface, error) {
 		return m.kubeClient, nil
 	}
 	return nil, args.Error(1)
+}
+
+func (m *MockClientGenerator) GatewayClient() (gateway.Interface, error) {
+	args := m.Called()
+	if args.Error(1) != nil {
+		return nil, args.Error(1)
+	}
+	m.gatewayClient = args.Get(0).(gateway.Interface)
+	return m.gatewayClient, nil
 }
 
 func (m *MockClientGenerator) IstioClient() (istioclient.Interface, error) {
@@ -115,7 +127,7 @@ func (suite *ByNamesTestSuite) TestAllInitialized() {
 			}: "TCPIngressesList",
 		}), nil)
 
-	sources, err := ByNames(mockClientGenerator, []string{"service", "ingress", "istio-gateway", "contour-httpproxy", "kong-tcpingress", "fake"}, minimalConfig)
+	sources, err := ByNames(context.TODO(), mockClientGenerator, []string{"service", "ingress", "istio-gateway", "contour-httpproxy", "kong-tcpingress", "fake"}, minimalConfig)
 	suite.NoError(err, "should not generate errors")
 	suite.Len(sources, 6, "should generate all six sources")
 }
@@ -124,7 +136,7 @@ func (suite *ByNamesTestSuite) TestOnlyFake() {
 	mockClientGenerator := new(MockClientGenerator)
 	mockClientGenerator.On("KubeClient").Return(fakeKube.NewSimpleClientset(), nil)
 
-	sources, err := ByNames(mockClientGenerator, []string{"fake"}, minimalConfig)
+	sources, err := ByNames(context.TODO(), mockClientGenerator, []string{"fake"}, minimalConfig)
 	suite.NoError(err, "should not generate errors")
 	suite.Len(sources, 1, "should generate fake source")
 	suite.Nil(mockClientGenerator.kubeClient, "client should not be created")
@@ -134,7 +146,7 @@ func (suite *ByNamesTestSuite) TestSourceNotFound() {
 	mockClientGenerator := new(MockClientGenerator)
 	mockClientGenerator.On("KubeClient").Return(fakeKube.NewSimpleClientset(), nil)
 
-	sources, err := ByNames(mockClientGenerator, []string{"foo"}, minimalConfig)
+	sources, err := ByNames(context.TODO(), mockClientGenerator, []string{"foo"}, minimalConfig)
 	suite.Equal(err, ErrSourceNotFound, "should return source not found")
 	suite.Len(sources, 0, "should not returns any source")
 }
@@ -143,16 +155,16 @@ func (suite *ByNamesTestSuite) TestKubeClientFails() {
 	mockClientGenerator := new(MockClientGenerator)
 	mockClientGenerator.On("KubeClient").Return(nil, errors.New("foo"))
 
-	_, err := ByNames(mockClientGenerator, []string{"service"}, minimalConfig)
+	_, err := ByNames(context.TODO(), mockClientGenerator, []string{"service"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"ingress"}, minimalConfig)
+	_, err = ByNames(context.TODO(), mockClientGenerator, []string{"ingress"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
+	_, err = ByNames(context.TODO(), mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"kong-tcpingress"}, minimalConfig)
+	_, err = ByNames(context.TODO(), mockClientGenerator, []string{"kong-tcpingress"}, minimalConfig)
 	suite.Error(err, "should return an error if kubernetes client cannot be created")
 }
 
@@ -162,10 +174,10 @@ func (suite *ByNamesTestSuite) TestIstioClientFails() {
 	mockClientGenerator.On("IstioClient").Return(nil, errors.New("foo"))
 	mockClientGenerator.On("DynamicKubernetesClient").Return(nil, errors.New("foo"))
 
-	_, err := ByNames(mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
+	_, err := ByNames(context.TODO(), mockClientGenerator, []string{"istio-gateway"}, minimalConfig)
 	suite.Error(err, "should return an error if istio client cannot be created")
 
-	_, err = ByNames(mockClientGenerator, []string{"contour-httpproxy"}, minimalConfig)
+	_, err = ByNames(context.TODO(), mockClientGenerator, []string{"contour-httpproxy"}, minimalConfig)
 	suite.Error(err, "should return an error if contour client cannot be created")
 }
 
